Rename user_plugins to plugin_library
