+++ CONTRIBUTING.md
@@ -21,7 +21,7 @@ This installs all required development dependencies including code quality tools
 - **`src/entity/config/`** - Configuration models and validation
 - **`src/entity/cli/`** - Developer CLI tools and utilities
 - **`src/entity/utils/`** - Shared utilities (logging, retries, etc.)
-- **`user_plugins/`** - Example plugins and local development plugins (not shipped with framework)
+- **`plugin_library/`** - Example plugins and local development plugins (not shipped with framework)
 
 ### Test Organization
 Mirror the source structure in tests:
@@ -86,8 +86,8 @@ pytest tests/performance/ -m benchmark
 
 ### Plugin Location Rules
 - **Built-in plugins:** Add to `src/entity/plugins/` if the plugin ships with the framework
-- **Example/demo plugins:** Add to `user_plugins/` for examples, local development, or sharing
-- **Custom plugins:** Users develop in `user_plugins/` or their own repositories
+- **Example/demo plugins:** Add to `plugin_library/` for examples, local development, or sharing
+- **Custom plugins:** Users develop in `plugin_library/` or their own repositories
 
 ### Plugin Architecture Rules
 - **Import restrictions:** Plugins must NOT import core modules directly (enforces architectural boundaries)
