+++ examples/intermediate_agent/README.md
@@ -1,6 +1,6 @@
 # Intermediate Agent
 
-The intermediate agent chains a prompt plugin with a custom responder from `user_plugins`. It stores a brief analysis during the THINK stage and outputs the final explanation in the OUTPUT stage.
+The intermediate agent chains a prompt plugin with a custom responder from `plugin_library`. It stores a brief analysis during the THINK stage and outputs the final explanation in the OUTPUT stage.
 
 ```bash
 poetry run python examples/intermediate_agent/main.py
