+++ examples/kitchen_sink/main.py
@@ -5,8 +5,8 @@ from pathlib import Path
 sys.path.insert(0, str(Path(__file__).parents[1] / "src"))
 
 from entity import agent, prompt
-from user_plugins.tools.calculator_tool import CalculatorTool
-from user_plugins.responders import ReactResponder
+from plugin_library.tools.calculator_tool import CalculatorTool
+from plugin_library.responders import ReactResponder
 
 
 @prompt
