+++ plugin_library/prompts/chain_of_thought.py
