+++ plugin_library/responders/complex_prompt_responder.py
