+++ tests/test_user_plugin_imports.py
@@ -2,8 +2,8 @@ import re
 from pathlib import Path
 
 
-def test_user_plugins_do_not_import_pipeline():
-    base = Path("user_plugins")
+def test_plugin_library_does_not_import_pipeline():
+    base = Path("plugin_library")
     pattern = re.compile(r"^(?:from|import)\s+(entity\.pipeline|pipeline)")
     offending = []
     for py_file in base.rglob("*.py"):
